requirements.txt
cffi==1.14.0
coverage==5.1
cryptography==2.9.2
mock==4.0.2
prettytable==0.7.2
pycparser==2.20
pynput==1.6.8
python-dateutil==2.8.1
python-xlib==0.27
PyYAML==5.3.1
six==1.14.0
